feat: force_native config (#31)
This is a bit of a hack but some airbyte sources are no longer listed as
enabled on pypi in the catalog of connectors but their packages are
still available on pypi. A lot of the low code/config only connectors
stopped being relaesed to pypi. This flag would force the wrapper to
attempt to use pypi whether its listed as available or not.